Align scheduler cycle widths with reference core

+++ app/src/main/cpp/pcsx2/pcsx2/Counters.h
@@ -60,13 +60,13 @@ struct Counter
 	};
 	u32 target, hold;
 	u32 rate, interrupt;
-	u64 startCycle;		// delta values should be signed.
+	u32 startCycle;		// delta values should be signed.
 };
 
 struct SyncCounter
 {
 	u32 Mode;
-	u64 startCycle;					// start cycle of timer
+	u32 startCycle;					// start cycle of timer
 	s32 deltaCycles;
 };
 
@@ -118,7 +118,7 @@ extern SyncCounter hsyncCounter;
 extern SyncCounter vsyncCounter;
 
 extern s32 nextDeltaCounter;		// delta until the next counter event (must be signed)
-extern u64 nextStartCounter;
+extern u32 nextStartCounter;
 extern uint g_FrameCount;
 
 extern void rcntUpdate_hScanline();
