+++ app/src/main/cpp/pcsx2/pcsx2/R3000AInterpreter.cpp
@@ -275,7 +275,7 @@ static s32 intExecuteBlock( s32 eeCycles )
 {
 	psxRegs.iopBreak = 0;
 	psxRegs.iopCycleEE = eeCycles;
-	u64 lastIOPCycle = 0;
+	u32 lastIOPCycle = 0;
 
 	while (psxRegs.iopCycleEE > 0)
 	{
